:construction: Started reworking the header-only part of the library into several translation units for compile time and memory improvements
